framework/backend/python/packages/k1s0-config/pyproject.toml
[project]
name = "k1s0-config"
version = "0.1.0"
description = "YAML-based configuration management for k1s0 Python services"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/k1s0_config"]
